drivers/adxl345: cleanup and fixes

+++ drivers/include/adxl345.h
@@ -156,7 +156,7 @@ typedef struct {
     uint8_t addr;                   /**< I2C address */
     adxl345_params_t *params;       /**< Device configuration */
     adxl345_interrupt_t interrupt;  /**< Interrupts configuration  */
-    float scale_factor;             /**< Scale factor for converting value to mg */
+    int16_t scale_factor;           /**< Scale factor for converting value to mg */
 } adxl345_t;
 
 /**
